# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_data_aggregation_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/data-aggregation-expected-revenue": {
            "get": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "8e279d3cf3da500ba81f4c7b1e0d8739",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "32e6f6578ca0ba3aa3d97837c4c9fdba",
                "requestBody": {
                    "request": "Create Expected Revenue",
                    "description": "Create Expected Revenue Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "short_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "mount": {
                                        "description": "0",
                                        "type": "text",
                                        "example": "Август"
                                    },
                                    "sum": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "company_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "17"
                                    },
                                    "year": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "company_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "short_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "mount": {
                                            "description": "0",
                                            "type": "text",
                                            "example": "Август"
                                        },
                                        "sum": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "company_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "17"
                                        },
                                        "year": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/data-aggregation-expected-revenue/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "d9c08262f38ce5445f4f8d855682d6b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/data-aggregation-expected-revenue/get-by-name/{company_name}": {
            "get": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "7b7ed99bb69ec77a16b9d2b2f1681f5c",
                "parameters": [
                    {
                        "name": "company_name",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/data-aggregation-expected-revenue/{id}": {
            "delete": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "74d906f2f2732bf5d51758f8ac934f08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Expected Revenue"
                ],
                "operationId": "0260651aa18d6497266d9e97dc2c9a0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update Expected Revenue",
                    "description": "Update Expected Revenue Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "company_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "short_name": {
                                        "description": "Название компании",
                                        "type": "text",
                                        "example": ""
                                    },
                                    "mount": {
                                        "description": "0",
                                        "type": "text",
                                        "example": "Август"
                                    },
                                    "sum": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "company_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "17"
                                    },
                                    "year": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "company_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "short_name": {
                                            "description": "Название компании",
                                            "type": "text",
                                            "example": ""
                                        },
                                        "mount": {
                                            "description": "0",
                                            "type": "text",
                                            "example": "Август"
                                        },
                                        "sum": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "company_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "17"
                                        },
                                        "year": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "2022"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                },
                                                "field2": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field2 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}